Refactor testbench.libcmd
