+++ spice/testbench.py
@@ -108,79 +108,16 @@ class testbench(testbench_common):
     # Generating eldo/spectre library inclusion string
     @property
     def libcmd(self):
-        """String
-        
-        Library inclusion string. Parsed from self.spicecorner -dictionary in
+        """str : Library inclusion string. Parsed from self.spicecorner -dictionary in
         the parent entity, as well as 'ELDOLIBFILE' or 'SPECTRELIBFILE' global
         variables in TheSDK.config.
         """
-        if not hasattr(self,'_libcmd'):
-            libfile = ""
-            corner = "top_tt"
-            temp = "27"
-            for optname,optval in self.parent.spicecorner.items():
-                if optname == "temp":
-                    temp = optval
-                if optname == "corner":
-                    corner = optval
-            if self.parent.model == 'eldo':
-                try:
-                    libfile = thesdk.GLOBALS['ELDOLIBFILE']
-                    if libfile == '':
-                        raise ValueError
-                    else:
-                        self._libcmd = "*** Eldo device models\n"
-                        self._libcmd += ".lib " + libfile + " " + corner + "\n"
-                except:
-                    self.print_log(type='W',msg='Global TheSDK variable ELDOLIBFILE not set.')
-                    self._libcmd = "*** Eldo device models (undefined)\n"
-                    self._libcmd += "*.lib " + libfile + " " + corner + "\n"
-                self._libcmd += ".temp " + str(temp) + "\n"
-            if self.parent.model == 'spectre':
-                try:
-                    libfile = thesdk.GLOBALS['SPECTRELIBFILE']
-                    if libfile == '':
-                        raise ValueError
-                    else:
-                        self._libcmd = "// Spectre device models\n"
-                        files = libfile.split(',')
-                        if len(files)>1:
-                            if isinstance(corner,list) and len(files) == len(corner):
-                                for path,corn in zip(files,corner):
-                                    if not isinstance(corn, list):
-                                        corn = [corn]
-                                    for c in corn:
-                                        self._libcmd += 'include "%s" section=%s\n' % (path,c)
-                            else:
-                                self.print_log(type='W',msg='Multiple entries in SPECTRELIBFILE but spicecorner wasn\'t a list or contained different number of elements!')
-                                self._libcmd += 'include "%s" section=%s\n' % (files[0], corner)
-                        else:
-                            self._libcmd += 'include "%s" section=%s\n' % (files[0], corner)
-                except:
-                    self.print_log(type='W',msg='Global TheSDK variable SPECTRELIBPATH not set.')
-                    self._libcmd = "// Spectre device models (undefined)\n"
-                    self._libcmd += "//include " + libfile + " " + corner + "\n"
-                self._libcmd += 'tempOption options temp=%s\n' % str(temp)
-            if self.parent.model == 'ngspice':
-                try:
-                    libfile = thesdk.GLOBALS['NGSPICELIBFILE']
-                    if libfile == '':
-                        raise ValueError
-                    else:
-                        self._libcmd = "*** Ngspice device models\n"
-                        self._libcmd += ".lib " + libfile + " " + corner + "\n"
-                except:
-                    self.print_log(type='W',msg='Global TheSDK variable ELDOLIBFILE not set.')
-                    self._libcmd = "*** Eldo device models (undefined)\n"
-                    self._libcmd += "*.lib " + libfile + " " + corner + "\n"
-                self._libcmd += ".temp " + str(temp) + "\n"
-        return self._libcmd
+
+        return self.testbench_simulator.libcmd
+
     @libcmd.setter
     def libcmd(self,value):
-        self._libcmd=value
-    @libcmd.deleter
-    def libcmd(self,value):
-        self._libcmd=None
+        self.testbench_simulator.libcmd = value
 
     # Generating netlist inclusion string
     @property
